bug fixing
